+++ docs/src/Box.doc.js
@@ -73,7 +73,7 @@ card(
       },
       {
         name: 'color',
-        type: `"blue" | "darkGray" | "darkWash" | "eggplant" | "gray" | "green" | "lightGray" | "lightWash" | "maroon" | "midnight" | "navy" | "olive" | "orange" | "orchid" | "pine" | "purple" | "red" | "transparent" | "watermelon" | "white"`,
+        type: `"blue" | "darkGray" | "darkWash" | "eggplant" | "gray" | "green" | "lightGray" | "lightWash" | "maroon" | "midnight" | "navy" | "olive" | "orange" | "orchid" | "pine" | "purple" | "red" | "transparent" | "transparentDarkGray" | "watermelon" | "white"`,
         defaultValue: 'transparent',
       },
       { name: 'fit', type: 'boolean', defaultValue: false },
@@ -512,11 +512,13 @@ card(
       'navy',
       'midnight',
       'purple',
+      'orchid',
       'eggplant',
       'maroon',
       'watermelon',
       'orange',
       'transparent',
+      'transparentDarkGray',
       'lightWash',
       'darkWash',
     ]}